refactor: 패키지 및 클래스 역할 변경
