+++ src/test/java/com/dowe/IntegrationTestSupport.java
@@ -5,7 +5,7 @@ import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.test.context.ActiveProfiles;
 
 import com.dowe.auth.application.OAuthProvider;
-import com.dowe.member.MemberRepository;
+import com.dowe.member.infrastructure.MemberRepository;
 
 @SpringBootTest
 @ActiveProfiles("test")
