+++ src/test/java/com/dowe/exception/GlobalExceptionHandlerTest.java
@@ -26,8 +26,8 @@ import com.dowe.auth.application.OAuthProvider;
 import com.dowe.auth.dto.LoginData;
 import com.dowe.auth.infrastructure.MemberTokenRepository;
 import com.dowe.member.Member;
-import com.dowe.member.MemberRepository;
 import com.dowe.member.Provider;
+import com.dowe.member.infrastructure.MemberRepository;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Charsets;
